chore(YALB-833): move grand hero folder to banner folder
