+++ components/02-molecules/banner/grand-hero/_yds-grand-hero.scss
@@ -1,0 +1,67 @@
+@use '../../../00-tokens/tokens';
+@use '../../../01-atoms/atoms';
+
+$break-grand-hero-banner: tokens.$break-m;
+
+.grand-hero-banner {
+  @include tokens.spacing-page-section($flush-top: true, $flush-bottom: true);
+
+  padding: 0;
+}
+
+.grand-hero-banner__content-wrapper {
+  position: relative;
+  overflow: hidden;
+  min-height: 30rem;
+  display: grid;
+  align-items: center;
+
+  [data-grand-hero-overlay-variation='contained'] & {
+    justify-content: start;
+  }
+}
+
+.grand-hero-banner__image {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+
+  img {
+    object-fit: cover;
+    height: 100%;
+    width: 100%;
+  }
+}
+
+.grand-hero-banner__content {
+  position: relative;
+  opacity: 0.6;
+
+  // prettier-ignore
+  padding: var(--size-spacing-5) var(--size-spacing-7);
+
+  [data-grand-hero-overlay-variation='contained'] & {
+    max-width: var(--size-component-layout-width-content);
+    margin: var(--size-spacing-site-gutter);
+  }
+
+  [data-grand-hero-overlay-variation='full'] & {
+    height: 100%;
+    display: flex;
+    align-items: center;
+  }
+}
+
+.grand-hero-banner__heading {
+  @include tokens.h2-yale-new;
+
+  margin-bottom: var(--size-spacing-2);
+}
+
+.grand-hero-banner__link {
+  @include atoms.link;
+
+  flex: 0 0 auto;
+}
