+++ components/02-molecules/banner/grand-hero/yds-grand-hero.twig
@@ -1,0 +1,58 @@
+{#
+ # Available Props:
+ # - grand_hero__content__background
+ # - grand_hero__overlay_variation: contained (default), full
+ #
+ # Available Variables:
+ # - grand_hero__heading
+ # - grand_hero__snippet
+ # - grand_hero__link__content
+ # - grand_hero__link__url
+ #
+ # Available Blocks:
+ # - grand_hero__media
+ #}
+
+{% set grand_hero__base_class = 'grand-hero-banner' %}
+{% set grand_hero__content__background = grand_hero__content__background|default('basic-black') %}
+{% set grand_hero__overlay_variation = grand_hero__overlay_variation|default('contained') %}
+
+{% set grand_hero__attributes = {
+  'data-grand-hero-overlay-variation': grand_hero__overlay_variation,
+  class: bem(grand_hero__base_class),
+} %}
+
+<div {{ add_attributes(grand_hero__attributes) }}>
+  <div {{ bem('content-wrapper', [], grand_hero__base_class) }}>
+    <div {{ bem('image', [], grand_hero__base_class) }}>
+      {% block grand_hero__image %}
+        {% include "@atoms/images/image/_responsive-image.twig" %}
+      {% endblock %}
+    </div>
+    <div {{ bem('content', [], grand_hero__base_class) }} data-component-theme={{ grand_hero__content__background }}>
+      <div {{ bem('content-inner', [], grand_hero__base_class) }}>
+        <div {{ bem('text', [], grand_hero__base_class) }}>
+          {% include "@atoms/typography/headings/yds-heading.twig" with {
+            heading__level: '1',
+            heading__blockname: grand_hero__base_class,
+            heading: banner__heading,
+          } %}
+          {% if banner__snippet %}
+            {% include "@atoms/typography/text/yds-text.twig" with {
+              text__base_class: 'snippet',
+              text__blockname: grand_hero__base_class,
+              text__content: banner__snippet,
+            } %}
+          {% endif %}
+        </div>
+        {% if banner__link__content and banner__link__url %}
+          {% include "@atoms/controls/text-link/yds-text-link.twig" with {
+            link__content: banner__link__content,
+            link__url: banner__link__url,
+            link__blockname: grand_hero__base_class,
+          } %}
+        {% endif %}
+      </div>
+    </div>
+  </div>
+</div>
