--- components/02-molecules/grand-hero/_yds-grand-hero.scss
@@ -1,43 +1,0 @@
-@use '../../00-tokens/tokens';
-@use '../../01-atoms/atoms';
-
-$break-grand-hero-banner: tokens.$break-m;
-
-.grand-hero-banner {
-  @include tokens.spacing-page-section($flush-top: true, $flush-bottom: true);
-
-  padding: 0;
-}
-
-.grand-hero-banner__image {
-  position: relative;
-
-  img {
-    object-fit: cover;
-    height: 100%;
-    width: 100%;
-  }
-}
-
-.grand-hero-banner__content {
-  width: 50%;
-  top: 205px;
-  left: 45px;
-  opacity: 0.6;
-  position: absolute;
-
-  // prettier-ignore
-  padding: var(--size-spacing-4) var(--size-spacing-7) var(--size-spacing-6);
-}
-
-.grand-hero-banner__heading {
-  @include tokens.h2-yale-new;
-
-  margin-bottom: var(--size-spacing-2);
-}
-
-.grand-hero-banner__link {
-  @include atoms.link;
-
-  flex: 0 0 auto;
-}
