--- components/02-molecules/grand-hero/yds-grand-hero.twig
@@ -1,50 +1,0 @@
-{#
- # Available Variables:
- # - grand_hero__heading
- # - grand_hero__snippet
- # - grand_hero__link__content
- # - grand_hero__link__url
- #
- # Available Blocks:
- # - grand_hero__media
- #}
-
-{% set grand_hero__base_class = 'grand-hero-banner' %}
-{% set grand_hero__content__background = grand_hero__content__background|default('basic-black') %}
-
-{% set grand_hero__attributes = {
-  class: bem(grand_hero__base_class),
-} %}
-
-<div {{ add_attributes(grand_hero__attributes) }}>
-  <div {{ bem('content-wrapper', [], grand_hero__base_class) }}>
-    <div {{ bem('image', [], grand_hero__base_class) }}>
-      {% block grand_hero__image %}
-        {% include "@atoms/images/image/_responsive-image.twig" %}
-      {% endblock %}
-    </div>
-    <div {{ bem('content', [], grand_hero__base_class) }} data-component-theme={{ grand_hero__content__background }}>
-      <div {{ bem('text', [], grand_hero__base_class) }}>
-        {% include "@atoms/typography/headings/yds-heading.twig" with {
-          heading__level: '1',
-          heading__blockname: grand_hero__base_class,
-          heading: grand_hero__heading,
-        } %}
-        {% if grand_hero__snippet %}
-          {% include "@atoms/typography/text/yds-text.twig" with {
-            text__base_class: 'snippet',
-            text__blockname: grand_hero__base_class,
-            text__content: grand_hero__snippet,
-          } %}
-        {% endif %}
-      </div>
-      {% if grand_hero__link__content and grand_hero__link__url %}
-        {% include "@atoms/controls/text-link/yds-text-link.twig" with {
-          link__content: grand_hero__link__content,
-          link__url: grand_hero__link__url,
-          link__blockname: grand_hero__base_class,
-        } %}
-      {% endif %}
-    </div>
-  </div>
-</div>
