+++ components/02-molecules/molecules.scss
@@ -20,4 +20,4 @@
 @forward './text/yds-text-field';
 @forward './text-with-image/yds-text-with-image';
 @forward './video/yds-video';
-@forward './grand-hero/yds-grand-hero';
+@forward './banner/grand-hero/yds-grand-hero';
